Remove NVM checksum check
There is a known issue with NIC: boot can hang when NIC tries to
write checksum. There is a  fix which removes it.

Hovewer, e1000e won't probe if checksum is invalid. This patch
removes this check. According discussion (see links in patch) it
is safe and can be used as work-around.